feat: ft_strlen from int to size_t

+++ ft_strlcpy.c
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	int		s;
+	size_t	s;
 
 	s = 0;
 	if (dstsize == 0)
